backend/requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
supabase
stripe
openai
anthropic
google-generativeai
resend
slowapi
python-dotenv
pytest
httpx
